+++ DriverProgram/Counter.cs
@@ -1,0 +1,12 @@
+﻿namespace DriverProgram
+{
+    public class Counter
+    {
+        public long Count { get; private set; }
+
+        public void Increment()
+        {
+            Count++;
+        }
+    }
+}
